initial commit
switchover is intended to be a function lib that can be sourced in different bash scripts. 

it provides 3 base functions : 1. sqlexec intended to execute sql statement and fetch resultats in a shell friendly way , 2. maxexec is intended to execute commands to maxscale and fetch results in a shell friendly way , 3. switchover is intended to execute all the necessary steps to switch one slavetoits new master.